SplunkSender: implement events buffering (disabled by default)
When enabled, when send_data() is called, events are put in a buffer.
There are no API calls made to Splunk unless the buffer overflows.
With this feature, the user is no longer required to manage its
own buffer, and also avoid performing an API calls for every event

This allows the user to perform a simple data processing loop:
def process_stuff():
	while get_some_event():
		splunk.send_data()
	splunk.flush_buffer()

A buffer-size of 2000 performs well in many use-cases. However, the feature
is kept disabled by default to avoid breaking change behavior.

If buffering is enabled, the user should call flush_buffer() before exit().
If not, then events still waiting in the buffer, if any, are lost.

Signed-off-by: Alexandre Bruyelles <git@jack.fr.eu.org>